requirements.txt
# Core dependencies
PySide6>=6.6.0
SQLAlchemy>=2.0.0
alembic>=1.13.0

# LLM integrations
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Environment and configuration
python-dotenv>=1.0.0

# Data handling
pydantic>=2.0.0

# Optional: Local LLM support
# ollama-python>=0.1.0  # Uncomment if using Ollama

# Optional: Audio support (for future STT/TTS)
# openai-whisper>=20231117  # Uncomment for local STT
# pyttsx3>=2.90  # Uncomment for local TTS

# Development dependencies
pytest>=7.4.0
pytest-qt>=4.2.0
black>=23.0.0
